Get Available Geographies

get_geographies either retrieves a data frame with all available geographies or all existing units for a given geography

get_geographies(geography = NULL)

Arguments

geography

geography as string or NULL

Value

a data.frame either with the columns Kurzname (geography name), ID (geography identifier) and NGeb (number of units) or with the columns Schlüssel (unit identifier) and Name (unit name).

Examples

 # Available geographies  
 get_geographies()
#>                       Kurzname  ID  NGeb
#> 1                         Bund  BU     1
#> 2                     West/Ost  WO     2
#> 3                Bundesländer   BL    16
#> 4                       Kreise KRE   401
#> 5             Gemeindeverbände GVB  4603
#> 6            Regierungsbezirke RBZ    31
#> 7                Kreisregionen  KR   361
#> 8         Raumordnungsregionen ROR    96
#> 9         Arbeitsmarktregionen AMR   223
#> 10            Braunkohlerevier BKR     5
#> 11                  Regionstyp RTI     3
#> 12              Regionstyp W/O RWO     6
#> 13           Regionstyp Länder RTL    32
#> 14                    Kreistyp KTI     4
#> 15                Kreistyp W/O KWO     8
#> 16             Kreistyp Länder KBL    48
#> 17          Städt./ländl. Raum SLI     2
#> 18      Städt./ländl. Raum W/O SWO     4
#> 19   Städt./ländl. Raum Länder SBL    28
#> 20       Raumtyp Lage (Kreise) BLK     4
#> 21   Wachsen-Schrumpfen Kreise WSK     5
#> 22         BBSR-Mittelbereiche  MB   878
#> 23           Großstadtregionen GSR    51
#> 24             Einzugsbereiche GRE     5
#> 25                 IHK-Bezirke IHK    80
#> 26          Stadt-/Gemeindetyp SGI     5
#> 27      Stadt-/Gemeindetyp W/O SGW    10
#> 28   Stadt-/Gemeindetyp Länder SGB    67
#> 29    Stadt-/Gemeindetyp diff. SGD     7
#> 30          Raumtyp Lage (GVB) BLG     4
#> 31     Wachsen-Schrumpfen Gem. WSG     6
#> 32                 RegioStaR17 R17    17
#> 33                  RegioStaR7  R7     7
#> 34                  RegioStaR2  R2     2
#> 35                 RegioStaRG5 RG5     5
#> 36                          EU  EU     2
#> 37                      NUTS 0  N0    28
#> 38                      NUTS 1  N1   104
#> 39                      NUTS 2  N2   281
#> 40                   Gemeinden GEM 11001
#> 41         Zentrale Orte (j/n) ZO2     2
#> 42        Zentrale Orte (zus.) ZO4     4
#> 43                 Oberzentren  OZ   153
#> 44               Mittelzentren  MZ   961
#> 45                Grundzentren  GZ  2514
#> 46          Kein zentraler Ort kZO  7377
#> 47 Europäische Metropolregion  EUM    14

  # All districts 
  get_geographies("KRS")
#> list()