Get Metadata for a Variable

get_metadata retrieves the metadata for a single variable

Usage

get_metadata(variable)

Arguments

variable

variable identifier as character string

Value

a data.frame with columns Name (variable name), Kurzname (short name), Algorithmus (information on measurement), Quelle (source) and Anmerkungen (notes).

Details

To get a list of available variables use get_variables.

See also

Examples

  # Metadata for GDP variable
 get_metadata("010")
#>                              Name    Kurzname Algorithmus
#> 1 Zahl der Arbeitslosen insgesamt Arbeitslose            
#>                                                                                       Quelle
#> 1 Laufende Raumbeobachtung des BBSR; Arbeitsmarktstatistik der Bundesagentur für Arbeit (BA)
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Anmerkungen
#> 1 Es handelt sich um die Zahl der Arbeitslosen im Jahresdurchschnitt. Arbeitslose (bzw. arbeitslose Arbeitssuchende) sind laut Definition der BA Personen, die vorübergehend nicht in einem Beschäftigungsverhältnis stehen, eine versicherungspflichtige, mindestens 15 Wochenstunden umfassende Beschäftigung suchen, arbeiten dürfen, arbeitsfähig und arbeitsbereit sind, in der Bundesrepublik wohnen, nicht jünger als 15 Jahre sind und die Altersgrenze für den Renteneintritt noch nicht erreicht haben und sich persönlich bei einer Agentur für Arbeit oder einem Jobcenter arbeitslos gemeldet haben.
#>   ID
#> 1 10