Get Available Themes

get_themes retrieves a data frame with all available themes

Usage

get_themes(geography)

Arguments

geography

geography as character string

Value

a data.frame with columns ID (theme identifier), Bereich (theme name) and Unterbereich (sub-theme name).

Details

To get a list of available geographies use get_geographies.

See also

Examples

 # Available themes  
 get_themes("KRE")
#>     ID                             Bereich
#> 1  000                       Absolutzahlen
#> 2  011                    Arbeitslosigkeit
#> 3  013                    Arbeitslosigkeit
#> 4  012                    Arbeitslosigkeit
#> 5  021                    Bauen und Wohnen
#> 6  022                    Bauen und Wohnen
#> 7  033  Beschäftigung und Erwerbstätigkeit
#> 8  032  Beschäftigung und Erwerbstätigkeit
#> 9  031  Beschäftigung und Erwerbstätigkeit
#> 10 034  Beschäftigung und Erwerbstätigkeit
#> 11 041                         Bevölkerung
#> 12 042                         Bevölkerung
#> 13 045                         Bevölkerung
#> 14 044                         Bevölkerung
#> 15 043                         Bevölkerung
#> 16 051                             Bildung
#> 17 052                             Bildung
#> 18 071           Flächennutzung und Umwelt
#> 19 072           Flächennutzung und Umwelt
#> 20 081 Medizinische und soziale Versorgung
#> 21 082 Medizinische und soziale Versorgung
#> 22 090                Öffentliche Finanzen
#> 23 060   Privateinkommen, Private Schulden
#> 24 100                 Raumwirksame Mittel
#> 25 160        SDG-Indikatoren für Kommunen
#> 26 110                   Siedlungsstruktur
#> 27 122                    Sozialleistungen
#> 28 121                    Sozialleistungen
#> 29 123                    Sozialleistungen
#> 30 131          Verkehr und Erreichbarkeit
#> 31 133          Verkehr und Erreichbarkeit
#> 32 132          Verkehr und Erreichbarkeit
#> 33 141                          Wirtschaft
#> 34 142                          Wirtschaft
#> 35 156            Zentrale Orte Monitoring
#> 36 155            Zentrale Orte Monitoring
#> 37 152            Zentrale Orte Monitoring
#> 38 151            Zentrale Orte Monitoring
#> 39 154            Zentrale Orte Monitoring
#> 40 153            Zentrale Orte Monitoring
#>                               Unterbereich
#> 1                            Absolutzahlen
#> 2                                Allgemein
#> 3                            Altersgruppen
#> 4                                 Struktur
#> 5            Baulandmarkt und Bautätigkeit
#> 6             Gebäude- und Wohnungsbestand
#> 7                  Atypische Beschäftigung
#> 8                            Qualifikation
#> 9                                 Struktur
#> 10           Wirtschafts- und Berufszweige
#> 11                          Altersstruktur
#> 12                    Bevölkerungsstruktur
#> 13                        Bundestagswahlen
#> 14       Natürliche Bevölkerungsbewegungen
#> 15                             Wanderungen
#> 16                      Ausbildungsangebot
#> 17                      Schulische Bildung
#> 18                          Flächennutzung
#> 19                                  Umwelt
#> 20                 Medizinische Versorgung
#> 21                      Soziale Versorgung
#> 22                    Öffentliche Finanzen
#> 23       Privateinkommen, Private Schulden
#> 24                     Raumwirksame Mittel
#> 25            SDG-Indikatoren für Kommunen
#> 26                       Siedlungsstruktur
#> 27                   Bedarfsgemeinschaften
#> 28                      Leistungsempfänger
#> 29                      Transferleistungen
#> 30                          Erreichbarkeit
#> 31                                 Pendler
#> 32                          Straßenverkehr
#> 33                          Fremdenverkehr
#> 34                Wirtschaftliche Leistung
#> 35 Bildungsversorgungs- und Kulturfunktion
#> 36          Gesundheitsversorgungsfunktion
#> 37         Öffentliche Verwaltungsfunktion
#> 38             Rahmendaten Zentraler Orte 
#> 39    Verkehrs- und Kommunikationsfunktion
#> 40   Wirtschafts- und Arbeitsmarktfunktion