Get Available Variables

get_variables retrieves a data frame with available variables

Usage

get_variables(theme, geography)

Arguments

theme

theme as character string

geography

geography as character string

Value

a data.frame with columns KurznamePlus (variable name), Bereich (theme identifier), Gruppe (variable identifier), BU, EU, Zeitreihe (all three unknown).

Details

To get a list of available geographies use get_geographies. To get a list of available themes use get_themes.

See also

Examples

 # Retrieves list of available variables in the  
 # theme "unemployment" for districts 
 get_variables(theme="011", geography="KRE")
#>               KurznamePlus Bereich Gruppe  BU EU Zeitreihe
#> 1        Arbeitslosenquote     011     12 KRE           24
#> 2 Arbeitslosenquote Frauen     011     13 KRE           15
#> 3 Arbeitslosenquote Männer     011     14 KRE           15
#> 4       Arbeitslose Frauen     011     15 KRE           28
#> 5       Arbeitslose Männer     011     16 KRE           28