Get Available Geographies

get_geographies either retrieves a data frame with all available geographies or all existing units for a given geography

Usage

get_geographies(geography = NULL)

Arguments

geography

geography as string or NULL

Value

a data.frame either with the columns Kurzname (geography name), ID (geography identifier) and NGeb (number of units) or with the columns Schlüssel (unit identifier) and Name (unit name).

Examples

# The SSL vertification seems to fail for some Linux systems. 
# This is likely because of a SSL certificate issue on the INKAR server. 
# Disabling SSL verification comes with risks: 
# https://curl.se/libcurl/c/CURLOPT_SSL_VERIFYPEER.html
httr::set_config(httr::config(ssl_verifypeer = 0L))

# Available geographies  
get_geographies()
#>                       Kurzname  ID  NGeb
#> 1                    Gemeinden GEM 10990
#> 2             Gemeindeverbände GVB  4600
#> 3                       Kreise KRE   400
#> 4                Kreisregionen  KR   361
#> 5                         Bund  BU     1
#> 6                     West-Ost  WO     2
#> 7                 Bundesländer  BL    16
#> 8           Regierungsbezirke  RBZ    31
#> 9         Raumordnungsregionen ROR    96
#> 10        Arbeitsmarktregionen AMR   223
#> 11         BBSR-Mittelbereiche  MB   878
#> 12                 IHK-Bezirke IHK    80
#> 13           Braunkohlereviere BKR     5
#> 14            Metropolregionen EUM    15
#> 15           Großstadtregionen GSR    51
#> 16             Einzugsbereiche GRE     5
#> 17                    Kreistyp KTI     4
#> 18              Kreistyp (W-O) KWO     8
#> 19             Kreistyp Länder KBL    48
#> 20                  Regionstyp RTI     3
#> 21            Regionstyp (W-O) RWO     6
#> 22         Regionstyp (Länder) RTL    32
#> 23          Städt./ländl. Raum SLI     2
#> 24    Städt./ländl. Raum (W-O) SWO     4
#> 25 Städt./ländl. Raum (Länder) SBL    28
#> 26          Stadt-/Gemeindetyp SGI     5
#> 27    Stadt-/Gemeindetyp (W-O) SGW    10
#> 28 Stadt-/Gemeindetyp (Länder) SGB    67
#> 29  Stadt-/Gemeindetyp (diff.) SGD     7
#> 30                  RegioStaR2  R2     2
#> 31                 RegioStaRG5 RG5     5
#> 32                  RegioStaR7  R7     7
#> 33                 RegioStaR17 R17    17
#> 34       Raumtyp Lage (Kreise) BLK     4
#> 35          Raumtyp Lage (GVB) BLG     4
#> 36         Zentrale Orte (j/n) ZO2     2
#> 37        Zentrale Orte (zus.) ZO4     4
#> 38                 Oberzentrum  OZ   153
#> 39               Mittelzentrum  MZ   961
#> 40                Grundzentrum  GZ  2514
#> 41          Kein zentraler Ort kZO  7377
#> 42                          EU  EU     1
#> 43                      NUTS 0  N0    28
#> 44                      NUTS 1  N1   104
#> 45                      NUTS 2  N2   281

# All districts 
get_geographies("KRS")
#> list()

# Reset configuration  
httr::reset_config()