Get Available Themes

get_themes retrieves a data frame with all available themes

get_themes(geography)

Arguments

geography

geography as character string

Value

a data.frame with columns ID (theme identifier), Bereich (theme name) and Unterbereich (sub-theme name).

Details

To get a list of available geographies use get_geographies.

See also

Examples

 # Available themes  
 get_themes("KRE")
#>     ID                             Bereich                      Unterbereich
#> 1  000                       Absolutzahlen                     Absolutzahlen
#> 2  011                    Arbeitslosigkeit                         Allgemein
#> 3  013                    Arbeitslosigkeit                     Altersgruppen
#> 4  012                    Arbeitslosigkeit                          Struktur
#> 5  021                    Bauen und Wohnen     Baulandmarkt und Bautätigkeit
#> 6  022                    Bauen und Wohnen      Gebäude- und Wohnungsbestand
#> 7  033  Beschäftigung und Erwerbstätigkeit           Atypische Beschäftigung
#> 8  032  Beschäftigung und Erwerbstätigkeit                     Qualifikation
#> 9  031  Beschäftigung und Erwerbstätigkeit                          Struktur
#> 10 034  Beschäftigung und Erwerbstätigkeit     Wirtschafts- und Berufszweige
#> 11 041                         Bevölkerung                    Altersstruktur
#> 12 043                         Bevölkerung              Bevölkerungsstruktur
#> 13 046                         Bevölkerung                  Bundestagswahlen
#> 14 044                         Bevölkerung                         Mobilität
#> 15 045                         Bevölkerung Natürliche Bevölkerungsbewegungen
#> 16 051                             Bildung                Ausbildungsangebot
#> 17 052                             Bildung                Schulische Bildung
#> 18 071           Flächennutzung und Umwelt                    Flächennutzung
#> 19 072           Flächennutzung und Umwelt                            Umwelt
#> 20 081 Medizinische und soziale Versorgung           Medizinische Versorgung
#> 21 082 Medizinische und soziale Versorgung                Soziale Versorgung
#> 22 090                Öffentliche Finanzen              Öffentliche Finanzen
#> 23 060   Privateinkommen, Private Schulden Privateinkommen, Private Schulden
#> 24 100                 Raumwirksame Mittel               Raumwirksame Mittel
#> 25 160        SDG-Indikatoren für Kommunen      SDG-Indikatoren für Kommunen
#> 26 110                   Siedlungsstruktur                 Siedlungsstruktur
#> 27 122                    Sozialleistungen             Bedarfsgemeinschaften
#> 28 121                    Sozialleistungen                Leistungsempfänger
#> 29 123                    Sozialleistungen                Transferleistungen
#> 30 131          Verkehr und Erreichbarkeit                    Erreichbarkeit
#> 31 133          Verkehr und Erreichbarkeit                           Pendler
#> 32 132          Verkehr und Erreichbarkeit                    Straßenverkehr
#> 33 141                          Wirtschaft                    Fremdenverkehr
#> 34 142                          Wirtschaft          Wirtschaftliche Leistung